adds median() features to initial long model. addresses issues #233 and #234.
Former-commit-id: 816fe74dce8d59efb0475792728946f19ce1c356 [formerly ab885c90d1f2a020f6a862e33f9a13bd74b77ba9] [formerly 3f3fbb2f3942c31583f99c2cf5f63b80679abc03 [formerly 5b3eb4bfe434ba4a78283f8f0e9c7971c5ecc66e]]
Former-commit-id: 5be3f7e5fd00b0cb00808847427858c47d286a23 [formerly 1415a91e8fdef871018b731228a012e99c1877e5]
Former-commit-id: 604ffa2125694738e39c52d8cb421575409b844a

+++ src/pop/nodevector.cc
@@ -248,8 +248,8 @@ namespace FT{
                         
                 }
                 /* cout << "valid terminals: "; */
-                /* for (const auto& i : ti) */ 
-                /*     cout << terminals[i]->name << "(" << terminals[i]->otype << ", " */ 
+                /* for (unsigned i = 0; i < ti.size(); ++i) */ 
+                /*     cout << terminals[ti[i]]->name << "(" << terminals[ti[i]]->otype << ", " */ 
                 /*          << tw[i] << "), "; */ 
                 /* cout << "\n"; */
                 
@@ -273,7 +273,7 @@ namespace FT{
                 // let fi be indices of functions whose output type matches otype and, if max_d==1,
                 // with no boolean inputs (assuming all input data is floating point) 
                 vector<size_t> fi;
-                vector<size_t> fw;  // function weights
+                vector<float> fw;  // function weights
                 bool fterms = in(term_types, 'f');   // are there floating terminals?
                 bool bterms = in(term_types, 'b');   // are there boolean terminals?
                 bool cterms = in(term_types, 'c');   // are there categorical terminals?
@@ -322,17 +322,21 @@ namespace FT{
                 assert(fi.size() > 0 && "The operator set specified results in incomplete programs.");
                 
                 // append a random choice from fs            
-                /* auto t = functions[r.random_choice(fi)]->rnd_clone(); */
-                //std::cout << t->name << " ";
-                /* cout << "choices: \n"; */
-                /* for (const auto& fis : fi) */
-                /*     cout << functions[fis]->name << "," ; */
+                /* cout << "function choices: \n"; */
+                /* for (unsigned fis =0; fis < fi.size(); ++fis) */
+                /*     cout << "(" << functions[fi[fis]]->name << "," << fw[fis] << ") ,"; */
                 /* cout << "\n"; */
+                
                 push_back(functions[r.random_choice(fi,fw)]->rnd_clone());
                 
                 /* std::cout << "back(): " << back()->name << "\n"; */ 
                 std::unique_ptr<Node> chosen(back()->clone());
                 /* std::cout << "chosen: " << chosen->name << "\n"; */ 
+                /* std::cout << "continue?" ; */
+                /* int cont = 0; */
+                /* std::cin >>  cont ; */ 
+                /* if (cont != 1) */
+                /*     HANDLE_ERROR_THROW("exiting"); */
                 // recurse to fulfill the arity of the chosen function
                 for (size_t i = 0; i < chosen->arity['f']; ++i)
                     make_tree(functions, terminals, max_d-1, term_weights, op_weights, 'f',
